be-api/pyproject.toml
[tool.poetry]
name = "chatbond-api"
version = "0.1.0"
description = ""
authors = ["Léo Audibert <leonard.audibert@gmail.com>"]
readme = "README.md"
packages = [{ include = "chatbond_api" }]

[tool.poetry.dependencies]
python = ">=3.8,<3.9"
setuptools = "^67.8.0"
pytz = "^2023.3"
django = "3.2"
django-configurations = "^2.4.1"
gunicorn = "^20.1.0"
newrelic = "^8.8.0"
djoser = "^2.2.0"
djangorestframework-simplejwt = "^5.2.2"
pandas = "^2.0.1"
django-extensions = "^3.2.1"
pgvector = "^0.1.8"
django-redis = "^5.2.0"
django-storages = "^1.13.2"
dramatiq = { extras = ["rabbitmq", "watch"], version = "^1.14.2" }
django-dramatiq = "^0.11.2"
requests = "^2.31.0"
django-apscheduler = "^0.6.2"
django-model-utils = "^4.3.1"
django-simple-history = "^3.3.0"
django-taggit = "^4.0.0"
djangorestframework = "3.12"
drf-spectacular = "^0.26.2"
django-cors-headers = "^4.0.0"
annoy = "1.17.2"
psycopg2-binary = "^2.9.6"
dj-database-url = "^2.0.0"
scikit-learn = "^1.2.2"
nltk = "^3.8.1"
sentencepiece = "^0.1.99"
huggingface-hub = "^0.14.1"
tqdm = "^4.65.0"
# Forced to use a diff version locally, otherwise poetry tries to install the cpu version by default.
torch = [
    { version = "2.0.0", markers = "sys_platform == 'darwin'", source = "pypi" },
    { version = "2.0.1+cpu", markers = "sys_platform != 'darwin'", source = "torchcpu1130" },
]
torchvision = [
    { version = "0.15.1", markers = "sys_platform == 'darwin'", source = "pypi" },
    { version = "0.15.2+cpu", markers = "sys_platform != 'darwin'", source = "torchcpu1130" },
]
transformers = ">=4.6.0,<5.0.0"
tokenizers = "^0.10.3"
scipy = "^1.10.1"
numpy = "^1.24.3"
# six = "^1.16.0" # TODO: I don't think I need that anymore, it was because I was trying to use django-extensions-shell alone
boto3 = "^1.26.141"
sentence-transformers = { git = "https://github.com/leomorpho/sentence-transformers#master" }
cent = "^4.1.0"
sentry-sdk = "^1.30.0"
django-ses = "^3.5.0"
django-anymail = { extras = ["amazon-ses"], version = "^10.1" }
redis = "^5.0.1"
pika = "^1.3.2"

[tool.poetry.group.dev.dependencies]
mypy = "^1.3.0"
ruff = "^0.0.269"
black = "^23.3.0"
pre-commit = "^3.3.2"
make = "^0.1.6.post2"
mock = "^5.0.2"
coverage = "^7.2.6"
parameterized = "^0.9.0"
faker = "^18.9.0"
ipython = ">=8.0,<8.13"
watchdog = "^3.0.0"

[build-system]
requires = ["poetry-core>=1.4.2"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "torchcpu1130"
url = "https://download.pytorch.org/whl/cpu/"
default = false
secondary = true
